requirements.txt
flask
python-dotenv
requests
sentence-transformers
nlp_engine
numpy
sklearn